Move concatenate into Bytes.
Delete ByteArrayUtils.
