+++ scandium-core/src/test/java/org/eclipse/californium/scandium/dtls/cipher/CCMBlockCipherTest.java
@@ -27,9 +27,9 @@ import javax.crypto.Cipher;
 import javax.crypto.SecretKey;
 import javax.crypto.spec.SecretKeySpec;
 
+import org.eclipse.californium.elements.util.Bytes;
 import org.eclipse.californium.scandium.category.Small;
 import org.eclipse.californium.scandium.dtls.ProtocolVersion;
-import org.eclipse.californium.scandium.util.ByteArrayUtils;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -48,7 +48,7 @@ public class CCMBlockCipherTest {
 	static final byte[] aesKeyBytes = new byte[]{(byte) 0xC9, 0x0E, 0x6A, (byte) 0xA2, (byte) 0xEF, 0x60, 0x34, (byte) 0x96,
 		(byte) 0x90, 0x54, (byte) 0xC4, (byte) 0x96, 0x65, (byte) 0xBA, 0x03, (byte) 0x9E};
 	static final SecretKey aesKey = new SecretKeySpec(aesKeyBytes, "AES");
-	static final SecretKey aesKey256 = new SecretKeySpec(ByteArrayUtils.concatenate(aesKeyBytes, aesKeyBytes), "AES");
+	static final SecretKey aesKey256 = new SecretKeySpec(Bytes.concatenate(aesKeyBytes, aesKeyBytes), "AES");
 
 	static boolean strongEncryptionAvailable;
 	
@@ -106,13 +106,13 @@ public class CCMBlockCipherTest {
 		
 		// additional data based on sequence number, type (APPLICATION DATA) and protocol version
 		additionalData = new byte[]{TYPE_APPL_DATA, (byte) protocolVer.getMajor(), (byte) protocolVer.getMinor(), 0, (byte) payloadLength};
-		additionalData = ByteArrayUtils.concatenate(seq_num, additionalData);
+		additionalData = Bytes.concatenate(seq_num, additionalData);
 		additionalData = adjustLength(additionalData, aLength);
 		// "explicit" part of nonce, intentionally different from seq_num which MAY be used as the explicit nonce
 		// but does not need to be used (at least that's my interpretation of the specs)
 		byte[] explicitNonce = new byte[]{1, 2, 3, 4, 5, 6, 7, 8};
 		// nonce used for encryption, "implicit" part + "explicit" part
-		nonce = ByteArrayUtils.concatenate(client_iv, explicitNonce);
+		nonce = Bytes.concatenate(client_iv, explicitNonce);
 		nonce = adjustLength(nonce, nonceLength);
 	}
